More changes

+++ sbpc.Timesheet.Data/Entity/Mileage.cs
@@ -1,13 +1,11 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace sbpc.Timesheet.Data.Entity
 {
     public class Mileage 
     {
         public int Id { get; set; }
-        public int mile { get; set; }
+        public float mile { get; set; }
         public string Note { get; set; }
         public DateTime Date { get; set; }
         public string EmployeeName { get; set; }
